Factor out CheckStudentWithLowGrades
